+++ scriptworker/test/test_task.py
@@ -374,7 +374,7 @@ async def test_reclaim_task_mock(context, mocker, proc):
     async def fake_reclaim(*args, **kwargs):
         return {'credentials': context.credentials}
 
-    async def fake_kill_task(*args):
+    async def fake_kill_proc(*args):
         count.append(args)
 
     def die(*args):
@@ -384,7 +384,7 @@ async def test_reclaim_task_mock(context, mocker, proc):
     context.temp_queue = mock.MagicMock()
     context.temp_queue.reclaimTask = fake_reclaim
     mocker.patch.object(pprint, 'pformat', new=die)
-    mocker.patch.object(swtask, 'kill_task', new=fake_kill_task)
+    mocker.patch.object(swproc, 'kill_proc', new=fake_kill_proc)
     await swtask.reclaim_task(context, context.task)
     if proc:
         assert len(count) == 1
@@ -401,7 +401,7 @@ async def test_max_timeout_noop(context, mocker):
     async def fake_kill(*args):
         called.append(args)
 
-    mocker.patch.object(swtask, 'kill_task', new=fake_kill)
+    mocker.patch.object(swtask, 'kill_proc', new=fake_kill)
     await swtask.max_timeout(context, "invalid_proc", 0)
     assert called == []
 
@@ -410,7 +410,7 @@ async def test_max_timeout_noop(context, mocker):
 async def test_max_timeout_mock(context, mocker):
 
 
-    async def fake_kill_task(_, msg, status):
+    async def fake_kill_proc(_, msg, status):
         assert msg == "Exceeded task_max_timeout of 1 seconds"
         assert status == context.config['task_max_timeout_status']
 
@@ -418,7 +418,7 @@ async def test_max_timeout_mock(context, mocker):
     proc.pid = 10000
     context.proc = proc
     context.config['task_max_timeout'] = 1
-    mocker.patch.object(swtask, "kill_task", new=fake_kill_task)
+    mocker.patch.object(swproc, "kill_proc", new=fake_kill_proc)
     await swtask.max_timeout(context, proc, 1)